Finsihed adding product to cart logic for both scenarios logged in users and guest users. Almost done with getting user's cart for whether a guest or logged in user (The current bug is with access token still valid after logging out). Finished merging guest cart after signing in if the user already had a cart

+++ server/src/services/cartService.ts
@@ -97,33 +97,44 @@ class CartService {
   }
 
   async mergeGuestCartIntoUserCart(guestCartId: string, userId: string) {
+    console.log("guestCartId => ", guestCartId);
     const guestCart = await this.cartRepository.findCartWithItemsById(
       guestCartId
     );
+    console.log("Guest cart => ", guestCart);
     if (!guestCart || !guestCart.cartItems.length) return;
 
-    const userCart = await this.getCart({ userId });
+    // Check if the user already has a cart
+    const userCart = await this.cartRepository.findCartByUserId(userId);
 
-    for (const item of guestCart.cartItems) {
-      const existingItem = await this.cartRepository.findCartItem(
-        userCart.id,
-        item.productId
-      );
-      if (existingItem) {
-        await this.cartRepository.updateCartItem(existingItem.id, {
-          quantity: existingItem.quantity + item.quantity,
-        });
-      } else {
-        await this.cartRepository.createCartItem({
-          cartId: userCart.id,
-          productId: item.productId,
-          quantity: item.quantity,
-        });
+    if (userCart) {
+      // User has an existing cart, merge items into it
+      for (const item of guestCart.cartItems) {
+        const existingItem = await this.cartRepository.findCartItem(
+          userCart.id,
+          item.productId
+        );
+        if (existingItem) {
+          await this.cartRepository.updateCartItem(existingItem.id, {
+            quantity: existingItem.quantity + item.quantity,
+          });
+        } else {
+          await this.cartRepository.createCartItem({
+            cartId: userCart.id,
+            productId: item.productId,
+            quantity: item.quantity,
+          });
+        }
       }
+      // Delete the guest cart since items are merged
+      await this.cartRepository.deleteCart(guestCartId);
+      return userCart;
+    } else {
+      // User has no cart, reassign the guest cart to them
+      console.log("updating guest cart with logged in user id => ", userId);
+      await this.cartRepository.updateCart(guestCartId, { userId });
+      return guestCart; // Return the reassigned cart
     }
-
-    await this.cartRepository.deleteCart(guestCartId);
-    return this.getCart({ userId });
   }
 }
